Added new analysis type: MasterMasterReplication
Previously qumomf calculated hash of the vshard configuration and compared
it for each replica and leader in the set. This calculations was based on json
encoding which led to false failovers in case when replica had another order
of the fields in the configuration.

Another problem was that qumomf do not know how to build a full vshard configuration
to end up to a consistent configuration in the replica set.

To eliminate this problems I added a new analysis type MasterMasterReplication which
qumomf can fix applying follower role to all masters except the leader.

Also I changed the hash calculation algorithm: not it deep sorts vshard config of the set
before calculating.